Minor refactoring and add proile page and add navigation bar
