+++ frontend/pages/login.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NextPage } from 'next';
-import LoginPage from '@/components/LandingPage/LoginPage';
+import LoginPage from '../components/LoginPage/LoginPage';
 
 const Login: NextPage = () => {
     return (
