+++ frontend/pages/profile.tsx
@@ -1,0 +1,11 @@
+import { NextPage } from 'next';
+
+const Profile: NextPage = () => {
+    return (
+        <div>
+            <h1>Profile Page</h1>
+        </div>
+    )
+}
+
+export default Profile;
